// projects/app-api/src/lib/api/models/licenseGenerateRequest.ts
/**
 * Dsd API
 * DSD Information
 *
 * The version of the OpenAPI document: 0.0.1-SNAPSHOT
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface LicenseGenerateRequest {
  email?: string;
  duration?: number;
}